DevBuilding 0.0.5 Add IHitable Add DummyDweller for test Add HitInfo +3 hours

+++ Assets/_Scripts/Shooter/Weapon/Weapon.cs
@@ -7,7 +7,7 @@ namespace Shooter
     public interface IWeapon
     {
         float ReloadTime { get; }
-        float Damage { get; }        
+        int Damage { get; }        
         float AttackDispersion { get; }
         float Range { get; }
         void Attack(Vector3 direction);
